% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bessel_bridge.R
\name{bb_max}
\alias{bb_max}
\title{Brownian bridge maximum point simulation}
\usage{
bb_max(x, y, s, t, a1, a2)
}
\arguments{
\item{x}{Start value of Brownian bridge}

\item{y}{End value of Brownian bridge}

\item{s}{Start time of Brownian bridge}

\item{t}{End time of Brownian bridge}

\item{a1}{Lower bound of mmaximum point (max(x,y) <= a1 < a2)}

\item{a2}{Upper bound of maximum point (max(x,y) <= a1 < a2)}
}
\value{
List with the simulated maxmimum, 'max', and time where minimum occurs, 'tau'
}
\description{
Simulates the maximum value of a Brownian bridge
}
\examples{
# simulate a maximum of a Brownian bridge between 0 and 1 in time [0,1]
bb_min(x=0, y=0, s=0, t=1, a1 = 0, a2 = 1)

}
