% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bessel_bridge.R
\name{bb_min}
\alias{bb_min}
\title{Brownian bridge minimum point simulatiion}
\usage{
bb_min(x, y, s, t, a1, a2)
}
\arguments{
\item{x}{Start value of Brownian bridge}

\item{y}{End value of Brownian bridge}

\item{s}{Start time of Brownian bridge}

\item{t}{End time of Brownian bridge}

\item{a1}{Lower bound of minimum point (a1 < a2 <= min(x,y))}

\item{a2}{Upper bound of minimum point (a1 < a2 <= min(x,y))}
}
\value{
List with the simulated minimum, 'min', and time where minimum occurs, 'tau'
}
\description{
Simulates the minimum value of a Brownian bridge
}
\examples{
# simulate a minimum of a Brownian bridge between 0 and 1 in time [0,1]
bb_min(x=0, y=0, s=0, t=1, a1 = -1, a2 = 0)

}
