% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bessel_layers.R
\name{calc_SdeltaK_1_intervals}
\alias{calc_SdeltaK_1_intervals}
\title{Calculate interval: [S^{delta,1}_{2k+1}, S^{delta,1}_{2k}]}
\usage{
calc_SdeltaK_1_intervals(x, y, s, t, min, v, k, previous_values = NULL)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}

\item{k}{integer value}

\item{previous_values}{defaults to NULL: but if k >= 2, then can pass the interval 
for (k-1), i.e. a vector of [S^{delta,1}_{2k-1}, S^{delta,1}_{2k-2}]}
}
\value{
vector of two values, S^{delta,1}_{2k+1} and S^{delta,1}_{2k}
}
\description{
This function calculates the interval [S^{delta,1}_{2k+1}, S^{delta, 1}_{2k}] (case where min(x,y) > min)
}
\examples{
calc_SdeltaK_1_intervals(x = 0, y = 0, s = 0, t = 1, min = -2, v = 1, k = 1)

}
