% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bessel_layers.R
\name{simulate_layered_brownian_bridge_bessel}
\alias{simulate_layered_brownian_bridge_bessel}
\title{Layered Brownian Bridge sampler}
\usage{
simulate_layered_brownian_bridge_bessel(x, y, s, t, a, l, sim_times)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{a}{vector/sequence of numbers}

\item{l}{integer number denoting Bessel layer, i.e. Brownian bridge is contained in [min(x,y)-a[l], max(x,y)+a[l]]}

\item{sim_times}{vector of real numbers to simulate Bessel bridge}
}
\value{
matrix of the simulated layered Brownian bridge path, first row is points X, second row are corresponding times
}
\description{
This function simulates a layered Brownian Bridge given a Bessel layer, at given times
}
\examples{
# simulate Bessel layer
bes_layer <- simulate_bessel_layer(x = 0, y = 0, s = 0, t = 1, a = seq(0.1, 1.0, 0.1))
# simulate layered Brownian bridge
simulate_layered_brownian_bridge_bessel(x = 0, y = 0, s = 0, t = 1, a = bes_layer$a, l = bes_layer$l, sim_times = seq(0.2, 0.8, 0.2))

}
